+++ src-ui/app/services/status-automations/status-change-general-events-automation.service.ts
@@ -76,27 +76,29 @@ export class StatusChangeGeneralEventsAutomationService {
       .subscribe(async ({ status, statusMessage, sleepMode }) => {
         const oldStatus = this.vrcUser?.status;
         const oldStatusMessage = this.vrcUser?.statusDescription;
-        await this.vrchat.setStatus(status, statusMessage);
-        if (await this.notifications.notificationTypeEnabled('AUTO_UPDATED_VRC_STATUS')) {
-          await this.notifications.send(
-            this.translate.instant('notifications.vrcStatusChanged.content', {
-              newStatus: (
-                (statusMessage ?? oldStatusMessage) +
-                ' (' +
-                (status ?? oldStatus) +
-                ')'
-              ).trim(),
-            })
-          );
+        const success = await this.vrchat.setStatus(status, statusMessage).catch((e) => false);
+        if (success) {
+          if (await this.notifications.notificationTypeEnabled('AUTO_UPDATED_VRC_STATUS')) {
+            await this.notifications.send(
+              this.translate.instant('notifications.vrcStatusChanged.content', {
+                newStatus: (
+                  (statusMessage ?? oldStatusMessage) +
+                  ' (' +
+                  (status ?? oldStatus) +
+                  ')'
+                ).trim(),
+              })
+            );
+          }
+          this.eventLog.logEvent({
+            type: 'statusChangedOnGeneralEvent',
+            reason: sleepMode ? 'SLEEP_MODE_ENABLED' : 'SLEEP_MODE_DISABLED',
+            newStatus: status,
+            oldStatus: oldStatus,
+            newStatusMessage: statusMessage,
+            oldStatusMessage: oldStatusMessage,
+          } as EventLogStatusChangedOnGeneralEvent);
         }
-        this.eventLog.logEvent({
-          type: 'statusChangedOnGeneralEvent',
-          reason: sleepMode ? 'SLEEP_MODE_ENABLED' : 'SLEEP_MODE_DISABLED',
-          newStatus: status,
-          oldStatus: oldStatus,
-          newStatusMessage: statusMessage,
-          oldStatusMessage: oldStatusMessage,
-        } as EventLogStatusChangedOnGeneralEvent);
       });
   }