+++ src-ui/app/services/vrchat.service.ts
@@ -237,7 +237,7 @@ export class VRChatService {
     info(`[VRChat] Logged in: ${this._user.value?.displayName}`);
   }
 
-  async setStatus(status: UserStatus | null, statusMessage: string | null): Promise<void> {
+  async setStatus(status: UserStatus | null, statusMessage: string | null): Promise<boolean> {
     // Throw if we don't have a current user
     const userId = this._user.value?.id;
     if (!userId) {
@@ -247,10 +247,11 @@ export class VRChatService {
     // Sanitize status message if needed
     statusMessage =
       statusMessage === null ? null : statusMessage.replace(/\s+/g, ' ').trim().slice(0, 32);
-    // Don't do anything if the status is not changing
-    if (status && this._user.value?.status === status) return;
-    // Don't do anything if the status message is not changing
-    if (statusMessage && this._user.value?.statusDescription === statusMessage) return;
+    const statusChange = status && this._user.value?.status !== status;
+    const statusMessageChange =
+      statusMessage && this._user.value?.statusDescription !== statusMessage;
+    // Don't do anything if there would be no changes
+    if (!statusChange && !statusMessageChange) return false;
     // Log status change
     if (status && statusMessage) {
       info(`[VRChat] Changing status to '${statusMessage}' ('${status}')`);
@@ -258,10 +259,7 @@ export class VRChatService {
       info(`[VRChat] Changing status to '${status}'`);
     } else if (statusMessage) {
       info(`[VRChat] Changing status message to '${statusMessage}'`);
-    } else {
-      return;
     }
-
     // Send status change request
     try {
       const body: Record<string, string> = {};
@@ -283,7 +281,9 @@ export class VRChatService {
       if (!result.result?.ok) throw result.result;
     } catch (e) {
       error(`[VRChat] Failed to update status: ${JSON.stringify(e)}`);
+      return false;
     }
+    return true;
   }
 
   public showLoginModal(autoLogin = false) {