+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Exposed HMD activity level over MQTT as "HMD On Head" (Experimental).
 
+### Added
+- Issue where the status message wasn't updated when the visibility remained the same
+
 ## [1.13.1]
 
 ### Fixed